Fix bugs of previous update

+++ velodyne_pointcloud/include/velodyne_pointcloud/point_types.h
@@ -28,9 +28,9 @@ namespace velodyne_pointcloud
   {
     PCL_ADD_POINT4D;                    // quad-word XYZ
     float    intensity;                 // laser intensity reading
-    uint16_t ring;                      // laser ring number
+    float ring;                         // laser ring number
     uint32_t time;                      // time stamp
-    uint8_t  firing;                    // firing number
+    float  firing;                      // firing number
     float    azimuth;                   // azimuth (deg)
     EIGEN_MAKE_ALIGNED_OPERATOR_NEW     // ensure proper alignment
   } EIGEN_ALIGN16;
