+++ velodyne_pointcloud/src/conversions/colors.cc
@@ -80,7 +80,7 @@ namespace velodyne_pointcloud
         p.z = inMsg->points[i].z;
 
         // color lasers with the rainbow array
-        int color = inMsg->points[i].ring % N_COLORS;
+        int color = (int)inMsg->points[i].ring % N_COLORS;
         p.rgb = *reinterpret_cast<float*>(rainbow+color);
 
         outMsg->points.push_back(p);
